requirements.txt
Flask==3.0.3
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
marshmallow==3.21.2
PyJWT==2.8.0
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
flask-mail==0.9.1